move network config files into a `networks` directory
