+++ src/config/networks/network.d.ts
